requirements.txt
# Core LangChain dependencies
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langchain-community==0.0.10

# API and HTTP requests
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For specific cricket data APIs
# sportmonks==1.0.0  # Uncomment if using Sportmonks SDK
# cricapi==1.0.0     # Uncomment if using CricAPI SDK
